# backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Bank API
  version: 1.0.0
  description: API documentation
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /history:
    get:
      summary: Get transaction history for account
      tags:
        - History
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: iban
          required: true
          schema:
            type: string
          description: IBAN of the account
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
          description: Start date (YYYY-MM-DD)
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
          description: End date (YYYY-MM-DD)
      responses:
        '200':
          description: List of transactions for the account in the given date range
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transactionId:
                      type: string
                    fromIban:
                      type: string
                    toIban:
                      type: string
                    amount:
                      type: number
                    currency:
                      type: string
                    description:
                      type: string
                    status:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /accounts:
    get:
      summary: Get accounts for current user
      tags:
        - Accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Accounts for the authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: 64a7b2f5e4b0c8a1d2f3g4h5
                  iban:
                    type: string
                    example: EX14114020040000300201311111
                  balance:
                    type: number
                    example: 1000.50
                  currency:
                    type: string
                    example: EUR
                  userId:
                    type: string
                    example: 64a7b2f5e4b0c8a1d2f3g4h5
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid request
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized access
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Account not found
...